+++ mayan/apps/common/forms.py
@@ -34,10 +34,14 @@ class ChoiceForm(forms.Form):
         self.fields['selection'].help_text = help_text
         self.fields['selection'].widget.disabled_choices = disabled_choices
         self.fields['selection'].widget.attrs.update(
-            {'size': 14, 'class': 'choice_form'}
+            {
+                'class': 'full-height', 'data-height-difference': '450'
+            }
         )
 
-    selection = forms.MultipleChoiceField(widget=DisableableSelectWidget())
+    selection = forms.MultipleChoiceField(
+        required=False, widget=DisableableSelectWidget()
+    )
 
 
 class FormOptions(object):